+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Deprecated
 
 ### Removed
+- src/utils function isArrayOfStrings is removed, use one liner .every(v => typeof v === "string") instead in the future.
 
 ### Fixed